# openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "BE JS Challenge",
      "version": "1.0.11"
    },
...
    "servers": [
      {
        "url": "https://localhost:5001"
      }
    ],
...
    "paths": {
...
      "/api/v1/users/{userId}": {
        "get": {
          "tags": [
            "user"
          ],
          "summary": "Find user by ID",
          "description": "Returns a user and its profile",
          "operationId": "getUserById",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "description": "ID of user to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "password": {
                            "type": "string"
                          }
                        }
                      },
                      "profile": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "identity_type": {
                            "type": "string"
                          },
                          "identity_number": {
                            "type": "string"
                          },
                          "address": {
                            "type": "string"
                          },
                          "user_id": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "user"
          ],
          "summary": "Updates a user and profile data",
          "description": "Update user and profile data by id",
          "operationId": "updateUser",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "description": "ID of user that needs to be updated",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "identity_number": {
                      "type": "string"
                    },
                    "identity_type": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "password": {
                            "type": "string"
                          }
                        }
                      },
                      "profile": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "identity_type": {
                            "type": "string"
                          },
                          "identity_number": {
                            "type": "string"
                          },
                          "address": {
                            "type": "string"
                          },
                          "user_id": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Client Error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "user"
          ],
          "summary": "Deletes a user",
          "description": "delete a user",
          "operationId": "deleteUser",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "description": "User id to delete",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "password": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/auth/login": {
        "post": {
          "tags": [
            "authentication"
          ],
          "summary": "Log in to user account",
          "description": "login a user",
          "operationId": "loginUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "object",
                        "properties": {
                          "token": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "403": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "User does not exist",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/auth/authenticate": {
        "get": {
          "tags": [
            "authentication"
          ],
          "description": "Authenticate token and return user profile",
          "operationId": "getProfile",
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "password": {
                            "type": "string"
                          }
                        }
                      },
                      "profile": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "identity_type": {
                            "type": "string"
                          },
                          "identity_number": {
                            "type": "string"
                          },
                          "address": {
                            "type": "string"
                          },
                          "user_id": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "403": {
              "description": "Token is not provided",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Unauthenticated",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }